toda la página reparacion hecha y empezando menú reparacion. Falta actualizar el stock tabla piezas y añadir piezas-reparacion para tener todo enlazado.

+++ TallerFitipaldiNuevo/Clases/Pieza.cs
@@ -16,14 +16,14 @@ namespace TallerFitipaldiNuevo.Clases
         public int Id { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
-        public float Precio { get; set; }
+        public decimal Precio { get; set; }
         public int Stock { get; set; }
 
         public Pieza()
         {
         }
 
-        public Pieza(string nombre, string descripcion, float precio, int stock)
+        public Pieza(string nombre, string descripcion, decimal precio, int stock)
             : this()
         {
             Nombre = nombre;
